Move domain to file domain/Shop
